// src/app/modules/portfolio/enums/EStacksAndTools.enum.ts
export enum EStacksAndTools {
  Angular = 1,
  Aws = 2,
  Bootstrap = 3,
  Csharp = 4,
  Css = 5,
  Docker = 6,
  DotNet = 7,
  DotNetCore = 8,
  Git = 9,
  GitActions = 10,
  GitHub = 11,
  GitLab = 12,
  Grafana = 13,
  Html = 14,
  Insomnia = 15,
  JavaScript = 16,
  Kafka = 17,
  MsSql = 18,
  MySql = 19,
  Oracle = 20,
  Postman = 21,
  Splunk = 22,
  TypeScript = 23,
  VisualStudio = 24,
  VisualStudioCode = 25,
  Yaml = 26,
  AngularJS = 27
}